restructured frontend to better match Flask patterns
