+++ .env-base
@@ -1,3 +1,4 @@
 MONGO_URL=mongodb://localhost:27017
 MONGO_VOLUME_PATH=./data
 PHOTO_DICT=../photos
+SECRET_KEY=secret
