+++ app/config/config.py
@@ -23,7 +23,9 @@ class Config:
         if self.photo_dict is None:
             raise RuntimeError('Missing environment variable: PHOTO_DICT')
 
-        self.app_secret_key = os.getenv('SECRET_KEY') or 'one-ring-to-rule-them-all'
+        self.app_secret_key = os.getenv('SECRET_KEY')
+        if self.app_secret_key is None:
+            raise RuntimeError('Missing environment variable: SECRET_KEY')
 
 
 def get_config() -> Config:
