+++ app/domain/training_data/blueprints/training_data_bp.py
@@ -1,0 +1,9 @@
+from flask import Blueprint
+
+training_data_bp = Blueprint(name='training_data_bp',
+                             import_name=__name__,
+                             url_prefix='/training_data',
+                             template_folder='../templates',
+                             static_folder='../static')
+
+from app.domain.training_data.routes import routes
