--- app/domain/training_data/controllers/controller.py
@@ -1,9 +1,0 @@
-from app.hello import app
-
-TRAINING_DATA_URL = '/training_data'
-
-@app.route(TRAINING_DATA_URL, methods=['POST'])
-def create_training_data():
-    pass
-
-# ...
