+++ app/domain/training_data/routes/routes.py
@@ -1,29 +1,18 @@
-from flask import Flask, render_template, request, redirect, jsonify, abort, url_for
-from bson import ObjectId
-import json
+from flask import render_template, request, redirect, jsonify, abort, url_for
 
-from app.config.config import get_config
+from app.domain.training_data.blueprints.training_data_bp import training_data_bp
 from app.domain.training_data.queries.create_command import CreateTrainingDataCommand
 from app.domain.training_data.services import training_data_service
 from app.domain.training_data.forms.training_data_form import TrainingDataForm
 
-app = Flask(__name__)
-app.secret_key = get_config().app_secret_key
 
-
-@app.route('/')
-@app.route('/index')
-def index():
-    return render_template('base.html')
-
-
-@app.route('/training_data')
+@training_data_bp.route('/all', methods=['GET'])
 def view_all_data():
     all_data = training_data_service.get_all().training_data
     return render_template('view_all_data.html', data=all_data)
 
 
-@app.route('/training_data/<_id>')
+@training_data_bp.route('/<_id>', methods=['GET'])
 def view_datapoint(_id):
     datapoint = training_data_service.get(_id)
     if datapoint is None:
@@ -32,10 +21,16 @@ def view_datapoint(_id):
         return render_template('view_single_datapoint.html', datapoint=datapoint)
 
 
-@app.route('/training_data/new', methods=['GET', 'POST'])
+@training_data_bp.route('/new', methods=['GET'])
 def publish_new_training_data():
     form = TrainingDataForm()
-    if request.method == 'POST' and form.validate_on_submit():
+    return render_template("training_data_form.html", form=form)
+
+
+@training_data_bp.route('/new', methods=['POST'])
+def handle_form():
+    form = TrainingDataForm()
+    if form.validate_on_submit():
         data = CreateTrainingDataCommand()
         data.set_total_area(form.total_area.data)
         data.set_total_length(form.total_length.data)
@@ -47,9 +42,4 @@ def publish_new_training_data():
         training_data_service.create(data, photo)
 
         return render_template('upload_success.html')
-
     return render_template("training_data_form.html", form=form)
-
-
-if __name__ == '__main__':
-    app.run()
