+++ app/domain/training_data/templates/view_single_datapoint.html
